refactor: move lead-scoring compute core under schemes/lead_scoring/ [LTV-Pf.1]
First half of the physical reorg (LTV-Pf). Relocates the lead-scoring
hidden-world compute core — simulation/, mechanisms/, structure/ — as whole
directories under leadforge/schemes/lead_scoring/, so the scheme owns its
internals. Hard break, no shims (decision D12): old import paths are removed
and every in-repo caller updated.

- git mv of the three directories (21 file renames; history preserved).
- Repo-wide rewrite of the three absolute import prefixes
  (leadforge.{simulation,mechanisms,structure} →
  leadforge.schemes.lead_scoring.{...}) across package, tests, and scripts;
  no residual bare references remain.
- Public API (leadforge.api, CLI) unchanged; package stays 1.x.
- CHANGELOG "Moved" note with the old→new path table; CLAUDE.md Repository Map
  updated; docs/ltv/roadmap.md + .agent-plan.md updated (Pf split into Pf.1
  compute core / Pf.2 render).

Scope: render/{snapshots,relational,tasks} and the lead-scoring schema specs
are NOT moved here — render relocation (with the relational.py split) is Pf.2;
the schema split lands with LTV-Pg. core.models / render.relational TYPE_CHECKING
hints now reference the new structure/simulation paths (a known core→scheme
type-only reference; layering cleanup tracked for M6).

Verified byte-identical to main (14/14 files of a pinned-timestamp bundle);
full suite 1508 passed / 51 skipped; ruff + mypy clean; scripts compile.
