+++ .agent-plan.md
@@ -40,12 +40,13 @@ early against the known-good lead-scoring path + physical reorg into
 `schemes/`**. (Framing follows Google `lifetime_value`/ZILN and Voyantis pLTV.)
 
 Status: `LTV-M0` landed (#102, #103, #106). `LTV-M1`: `LTV-Pb` merged (#104);
-`LTV-Pc` (pLTV feature/task specs) still outstanding. `LTV-M2`: `LTV-Pd`
-(scheme protocol + registry) merged (#107); `LTV-Pe` (scheme owns bundle
-rendering — second half of the seam) opened as **#108** (awaiting review,
-verified byte-identical). M2 reordered so the render seam precedes the physical
-move. Next in M2: `LTV-Pf` (physically move lead-scoring pipeline into
-`schemes/lead_scoring/`), then `LTV-Pg` (scaffold `schemes/lifecycle/`).
+`LTV-Pc` (pLTV feature/task specs) still outstanding. `LTV-M2`: `LTV-Pd` (#107)
+and `LTV-Pe` (#108) merged (scheme protocol + render seam). `LTV-Pf` (physical
+move, **hard break / no shims** per D12) split into Pf.1 (compute core —
+simulation/mechanisms/structure moved) opened as **#NNN**, and Pf.2 (render
+move, pending). Verified byte-identical. Sibling `leadforge-datasets-private`
+build scripts must update to the new import paths (breakage issue filed). Next:
+`LTV-Pf.2` (render), then `LTV-Pg` (scaffold `schemes/lifecycle/`).
 
 ---
