+++ scripts/spike_category_signal.py
@@ -27,9 +27,9 @@ sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 from leadforge.api.generator import Generator
 from leadforge.core.rng import RNGRoot
 from leadforge.render.snapshots import build_snapshot
-from leadforge.simulation.engine import simulate_world
-from leadforge.simulation.population import PopulationResult, build_population
-from leadforge.structure.sampler import sample_hidden_graph
+from leadforge.schemes.lead_scoring.simulation.engine import simulate_world
+from leadforge.schemes.lead_scoring.simulation.population import PopulationResult, build_population
+from leadforge.schemes.lead_scoring.structure.sampler import sample_hidden_graph
 
 SEED = 42
 N_LEADS = 5000
