+++ tests/render/test_render.py
@@ -10,9 +10,9 @@ import pytest
 from leadforge.core.models import GenerationConfig
 from leadforge.core.rng import RNGRoot
 from leadforge.schema.features import LEAD_SNAPSHOT_FEATURES
-from leadforge.simulation.engine import simulate_world
-from leadforge.simulation.population import build_population
-from leadforge.structure.sampler import sample_hidden_graph
+from leadforge.schemes.lead_scoring.simulation.engine import simulate_world
+from leadforge.schemes.lead_scoring.simulation.population import build_population
+from leadforge.schemes.lead_scoring.structure.sampler import sample_hidden_graph
 
 # ---------------------------------------------------------------------------
 # Shared fixtures
