+++ tests/render/test_snapshot_windowed.py
@@ -12,9 +12,9 @@ import pytest
 from leadforge.core.models import GenerationConfig
 from leadforge.core.rng import RNGRoot
 from leadforge.render.snapshots import build_snapshot
-from leadforge.simulation.engine import simulate_world
-from leadforge.simulation.population import build_population
-from leadforge.structure.sampler import sample_hidden_graph
+from leadforge.schemes.lead_scoring.simulation.engine import simulate_world
+from leadforge.schemes.lead_scoring.simulation.population import build_population
+from leadforge.schemes.lead_scoring.structure.sampler import sample_hidden_graph
 
 
 def _make_narrative(seed: int = 42):
